+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Rage damage bonus fix.
 - Fix issue with importing items when updating them.
+- Fix an issues with some Save DC parsing in feature text.
+- Word Tree Barbarian features.
 
 # 6.0.13