+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next Up
+# 6.0.14
 
 - Rage damage bonus fix.
 - Fix issue with importing items when updating them.